Move functionality into projectmanager to allow closing editors when the document has been removed by undo
