+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/ProjectManager.java
@@ -26,9 +26,12 @@ import java.lang.reflect.InvocationTargetException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.LinkedList;
 import java.util.List;
+import java.util.ListIterator;
 import java.util.Optional;
 import java.util.Set;
+import java.util.Stack;
 import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
@@ -37,7 +40,7 @@ import org.corpus_tools.hexatomic.core.errors.ErrorService;
 import org.corpus_tools.hexatomic.core.events.salt.SaltNotificationFactory;
 import org.corpus_tools.hexatomic.core.handlers.OpenSaltDocumentHandler;
 import org.corpus_tools.hexatomic.core.undo.ChangeSet;
-import org.corpus_tools.hexatomic.core.undo.UndoManager;
+import org.corpus_tools.hexatomic.core.undo.ReversibleOperation;
 import org.corpus_tools.salt.SaltFactory;
 import org.corpus_tools.salt.common.SDocument;
 import org.corpus_tools.salt.common.SDocumentGraph;
@@ -95,10 +98,10 @@ public class ProjectManager {
   @Inject
   SaltNotificationFactory notificationFactory;
 
-  @Inject
-  UndoManager undoManager;
-
   private boolean hasUnsavedChanges;
+
+  private final Stack<ChangeSet> changeSets = new Stack<>();
+  private final List<ReversibleOperation> uncommittedChanges = new LinkedList<>();
 
   @PostConstruct
   void postConstruct() {
@@ -284,6 +287,7 @@ public class ProjectManager {
       // Ignore, this can happen when the application is closing
     }
   }
+
 
   /**
    * Saves all documents and the corpus structure as Salt project to a new location.
@@ -505,9 +509,94 @@ public class ProjectManager {
    * Adds a checkpoint. A user will be able to undo all changes made between checkpoints.
    */
   public void addCheckpoint() {
-    ChangeSet changes = undoManager.commitCanges();
+    ChangeSet changes = null;
+    if (!uncommittedChanges.isEmpty()) {
+      // All recorded changes are reversable without saving and loading the whole document, create
+      // a different kind of checkpoint using a copy of all uncomitted changes
+      log.debug("Adding {} changes as checkpoint to undo list", uncommittedChanges.size());
+      changes = new ChangeSet(uncommittedChanges);
+      changeSets.add(changes);
+    }
+
+    // All uncommitted changes have been handled: restore internal recored state to default:
+    uncommittedChanges.clear();
+
     if (changes != null) {
       events.send(Topics.ANNOTATION_CHECKPOINT_CREATED, changes);
+    }
+  }
+
+  /**
+   * Checks whether it is possible to perform an undo.
+   * 
+   * @return True if undo is possible.
+   */
+  public boolean canUndo() {
+    return !changeSets.isEmpty();
+  }
+
+  /**
+   * Undoes all changes made in the last checkpoint
+   */
+  public void undo() {
+
+    // Make sure the last uncommited changes are added as a checkoint, but don't fire an event
+    if (!uncommittedChanges.isEmpty()) {
+      log.warn("Adding checkpoint for {} uncommited changes before performing undo.",
+          uncommittedChanges.size());
+      changeSets.add(new ChangeSet(uncommittedChanges));
+      uncommittedChanges.clear();
+    }
+
+    if (canUndo()) {
+      notificationFactory.setSuppressingEvents(true);
+      // Restore all documents from the last checkpoint
+      ChangeSet lastChangeSet = this.changeSets.pop();
+      // Iterate over the elements in reversed order
+      ListIterator<ReversibleOperation> itLastChangeSet =
+          lastChangeSet.getChanges().listIterator(lastChangeSet.getChanges().size());
+      while (itLastChangeSet.hasPrevious()) {
+        ReversibleOperation op = itLastChangeSet.previous();
+        op.restore();
+      }
+      notificationFactory.setSuppressingEvents(false);
+
+      // Performing an undo might unload documents, close the editors to avoid errors
+      // and give the user a visually clue (s)he unloaded/removed a document.
+      closeEditorsForRemovedDocuments();
+
+      events.send(Topics.ANNOTATION_CHECKPOINT_RESTORED, lastChangeSet);
+
+      // TODO: how to implement redo?
+    }
+  }
+
+  @Inject
+  @org.eclipse.e4.core.di.annotations.Optional
+  private void subscribeUndoOperationAdded(
+      @UIEventTopic(Topics.UNDO_OPERATION_ADDED) Object element) {
+    if (element instanceof ReversibleOperation) {
+      uncommittedChanges.add((ReversibleOperation) element);
+    }
+  }
+
+  /**
+   * Close all open editors where the document has been unloaded or removed by the undo action.
+   */
+  private void closeEditorsForRemovedDocuments() {
+
+    try {
+      for (MPart part : partService.getParts()) {
+        String docID = part.getPersistedState().get(OpenSaltDocumentHandler.DOCUMENT_ID);
+        if (docID != null && !docID.isEmpty()) {
+          Optional<SDocument> document = this.getDocument(docID);
+          if (!document.isPresent() || document.get().getDocumentGraph() == null) {
+            partService.hidePart(part);
+          }
+        }
+      }
+    } catch (IllegalStateException ex) {
+      // Ignore, this can happen when the application is closing
     }
   }
