+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/handlers/UndoHandler.java
@@ -1,23 +1,23 @@
 package org.corpus_tools.hexatomic.core.handlers;
 
 import javax.inject.Inject;
-import org.corpus_tools.hexatomic.core.undo.UndoManager;
+import org.corpus_tools.hexatomic.core.ProjectManager;
 import org.eclipse.e4.core.di.annotations.CanExecute;
 import org.eclipse.e4.core.di.annotations.Execute;
 
 public class UndoHandler {
 
   @Inject
-  private UndoManager undoManager;
+  private ProjectManager projectManager;
 
   @Execute
   public void execute() {
-    undoManager.undo();
+    projectManager.undo();
   }
 
   @CanExecute
   public boolean canExecute() {
-    return undoManager.canUndo();
+    return projectManager.canUndo();
   }
 
 }
